fix: resolved testing of client components
Resolved the issue of testing client components that use hooks like useState, **temporarily**. This approach still produces a warning when executing tests, but works for now:

```
Warning: Render methods should be a pure function of props and state; triggering nested component updates from render is not allowed. If necessary, trigger nested updates in componentDidUpdate.

Check the render method of TestComponent.
```