ci: added mergify rule to validate PR base
Whenever a Pull requests base branch is set to the main branch but the head is different from the canary branch, then the author is notified that the base branch must be changed. The reason for this is that all changes must first be merged onto the canary branch, before the canary branch is merged into the main branch.