ci: changed auto-assign author rule in mergify.yml
The reason for this change is that previously the author of a PR was always assigned to it. Given that now PRs are automatically created via GiHub Actions using a PTA of Marty-Byrde, Marty-Byrde is always automatically the author of a PR. By changing the rule, Marty-Byrde will only be assigned when he is the sole-author.