ref: added space-holder to ColorModeSwitcher
When no colormode is defined, then the ColorModeSwitcher will not render any element until it is defined. This can lead to weird shifts in the layout when the mode changes from undefined to light or dark.

By adding a space-holder the ColorModeSwitcher will occupy its space even when no mode is yet set.