feat: added auto-close logic to OpenCloseButton component
Previously, the mobile sidebar remained open after clicking navigation-items, which felt unnatural and was confusing. By adding this logic the mobile sidebar is automatically closed, when the route changes, thus when a navigation item is clicked.

The logic was added inside the OpenCloseButton, because it shares the same agenda of closing the mobile sidebar.