ci: added continue-on-error to push docker-image
The reason for this change is that a push on the main branch may not lead to a new release, which would mean that this workflow tries to push a new docker-image with an existing image-tag to the registry, which will throw an error. To prevent this from failing the workflow, this argument was added to the push-step.