ci: updated latest tag retrieval in auto-create-pr-canary
The reason for this change is that previously the tag-retrieval returned no tag, because the checkout did not collect all the tags. This issue has been resolved. Also pre-releases of the canary branch are discarded, because we want to have the latest tag of the main branch => latest release to display it in the PR title.