ci: added mergify rule to auto-update canary when main is ahead
This rule was previously removed as it was updated to auto-update branches that want to merge into canary. However, this has lead to the canary branches being un-mergeable, because they are behind and updating the branch has to be done manually, which requires the protection rules to be disabled temporarily. By adding this pull request rule to mergify, this issue should be resolved.